refactor; moved warnings into package
